building multiple comments format
