+++ react-vite/src/components/BookPage/BookPage_original.css
@@ -1,0 +1,158 @@
+.book-opener {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+}
+
+.book-preview > img {
+  object-fit: contain;
+  max-width: calc(100vw / 2);
+  max-height: calc(100vh - 51px);
+}
+.book-preview {
+  margin: auto;
+}
+
+.book-opener {
+  overflow-y: hidden;
+}
+
+.book-title-author {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  border-right: 2px solid rgba(0, 0, 0, 0.119);
+}
+
+.book-content {
+  font-size: x-large;
+  word-wrap: break-word;
+  white-space: pre-wrap;
+  font-family: "Times New Roman", Times, serif;
+  line-height: 40px;
+  overflow-y: auto;
+  max-height: calc(100vh - 100px);
+  padding: 1em 20vw;
+  margin: 0px;
+}
+
+.comment {
+  border-bottom: 3px double rgba(200, 150, 13, 0.693);
+  cursor: pointer;
+  padding-bottom: 5px;
+  position: static;
+}
+
+.comment-content {
+  position: absolute;
+  display: flex;
+  border: 1px solid black;
+  background-color: rgb(250, 236, 227);
+  border-radius: 5px;
+  padding: 5px;
+  font-size: medium;
+  margin-left: 10px;
+  width: 350px;
+  height: 175px;
+  bottom: 1px;
+  left: 36%;
+}
+
+.user-comment {
+  border-bottom: 3px double rgba(29, 13, 200, 0.693);
+}
+
+.close-comment {
+  position: absolute;
+  right: 10px;
+  cursor: pointer;
+  font-size: larger;
+}
+
+.delete-comment {
+  position: absolute;
+  left: 18px;
+  top: 90px;
+  cursor: pointer;
+  font-size: larger;
+  color: rgb(255, 131, 131);
+}
+
+.comment-avatar {
+  width: 50px;
+  height: 50px;
+  margin: 0px 0px;
+  border-radius: 40px;
+  object-fit: contain;
+  margin-right: 5px;
+  cursor: pointer;
+}
+.comment-no-avatar {
+  background-color: rgb(249, 152, 87);
+  width: 35px;
+  height: 35px;
+  margin: 7px 0px;
+  border-radius: 40px;
+  object-fit: contain;
+  margin-right: 5px;
+  font-size: small;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-weight: bold;
+  cursor: pointer;
+}
+.comment-user {
+  font-weight: bold;
+  padding-top: 16px;
+  height: 20px;
+}
+
+.comment-message {
+  overflow-y: auto;
+  word-wrap: break-word;
+  height: 80px;
+  width: 290px;
+  font-family: "Times New Roman", Times, serif;
+  font-size: medium;
+}
+
+.edit-message {
+  resize: none;
+  background-color: transparent;
+  border: none;
+  cursor: pointer;
+}
+.comment-reactions {
+  height: 50px;
+  width: 290px;
+  display: flex;
+  align-items: center;
+  flex-wrap: wrap;
+  overflow-y: auto;
+}
+
+.add-emoji {
+  position: absolute;
+  bottom: 12px;
+  left: 12px;
+  font-size: x-small;
+  font-weight: bold;
+}
+
+.add-reaction {
+  position: absolute;
+  height: 20px;
+  width: 20px;
+  cursor: pointer;
+  color: gray;
+  left: 20px;
+  bottom: 20px;
+}
+
+.reaction-display {
+  border: 1px solid gray;
+  border-radius: 8px;
+  margin: 3px;
+  width: 40px;
+}
